Osciloscope implementation

+++ android/mpu9250/app/src/main/java/jp/araobp/mpu9250/Mpu9250Data.kt
@@ -2,10 +2,10 @@ package jp.araobp.mpu9250
 
 data class Mpu9250Data (
     val seq: UInt,
-    val ax: Float,
-    val ay: Float,
-    val az: Float,
-    val gx: Float,
-    val gy: Float,
-    val gz: Float
+    val ax: Short,
+    val ay: Short,
+    val az: Short,
+    val gx: Short,
+    val gy: Short,
+    val gz: Short
 )
